패키지 리팩토링
- 목적에 따른 패키지 리팩토링
  - 사용자(user)
  - 계좌/계좌내역(account)
  - 카테고리(category)
