+++ app/src/main/java/com/codesoom/project/core/application/AccountService.java
@@ -3,7 +3,7 @@ package com.codesoom.project.core.application;
 import com.codesoom.project.common.error.AccountNotFoundException;
 import com.codesoom.project.core.domain.Account;
 import com.codesoom.project.core.domain.AccountRepository;
-import com.codesoom.project.web.dto.AccountCreationData;
+import com.codesoom.project.web.dto.account.AccountCreationData;
 import com.github.dozermapper.core.Mapper;
 import org.springframework.stereotype.Service;
