+++ app/src/main/java/com/codesoom/project/core/application/CategoryService.java
@@ -2,7 +2,7 @@ package com.codesoom.project.core.application;
 
 import com.codesoom.project.core.domain.Category;
 import com.codesoom.project.core.domain.CategoryRepository;
-import com.codesoom.project.web.dto.CategoryRegistrationData;
+import com.codesoom.project.web.dto.category.CategoryRegistrationData;
 import com.github.dozermapper.core.Mapper;
 import org.springframework.stereotype.Service;
