+++ app/src/main/java/com/codesoom/project/web/dto/account/AccountCreationData.java
@@ -1,20 +1,18 @@
-package com.codesoom.project.web.dto;
+package com.codesoom.project.web.dto.account;
 
 import com.github.dozermapper.core.Mapping;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 
 @Getter
 @Builder
+@NoArgsConstructor
 @AllArgsConstructor
-public class UserRegistrationData {
-    @NotBlank
-    @Mapping("email")
-    private String email;
-
+public class AccountCreationData {
     @NotBlank
     @Mapping("name")
     private String name;
