+++ app/src/main/java/com/codesoom/project/web/dto/category/CategoryResultData.java
@@ -1,4 +1,4 @@
-package com.codesoom.project.web.dto;
+package com.codesoom.project.web.dto.category;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
