+++ app/src/test/java/com/codesoom/project/core/application/CategoryServiceTest.java
@@ -2,7 +2,7 @@ package com.codesoom.project.core.application;
 
 import com.codesoom.project.core.domain.Category;
 import com.codesoom.project.core.domain.CategoryRepository;
-import com.codesoom.project.web.dto.CategoryRegistrationData;
+import com.codesoom.project.web.dto.category.CategoryRegistrationData;
 import com.github.dozermapper.core.DozerBeanMapperBuilder;
 import com.github.dozermapper.core.Mapper;
 import org.junit.jupiter.api.BeforeEach;
